Dissocaiting Shop from ShopStage
Stage logic separated from shop logic
